Moved taptools code to its own module.
